Improved the structure of the project
